app/requirements.txt
requests
discord.py
geopy
requests-random-user-agent
pyyaml
termcolor
telethon